+++ Minecraft.Client/Common/DLC/DLCSkinFile.cpp
@@ -79,7 +79,7 @@ void DLCSkinFile::addParameter(DLCManager::EDLCParameterType type, const wstring
 				{
 					i++;
 				}
-				int iLast=static_cast<int>(creditValue.find_last_of(L" ", i));
+				size_t iLast=creditValue.find_last_of(L" ", i);
 				switch(XGetLanguage())
 				{
 				case XC_LANGUAGE_JAPANESE:
@@ -88,7 +88,7 @@ void DLCSkinFile::addParameter(DLCManager::EDLCParameterType type, const wstring
 					iLast = maximumChars;
 					break;
 				default:
-					iLast=static_cast<int>(creditValue.find_last_of(L" ", i));
+					iLast=creditValue.find_last_of(L" ", i);
 					break;
 				}
