+++ Minecraft.Client/Orbis/Network/SQRNetworkManager_Orbis.cpp
@@ -685,7 +685,7 @@ bool SQRNetworkManager_Orbis::FriendRoomManagerSearch()
 	}
 
 	// Free up any external data that we received from the previous search
-	for( int i = 0; i < m_aFriendSearchResults.size(); i++ )
+	for( size_t i = 0; i < m_aFriendSearchResults.size(); i++ )
 	{
 		if(m_aFriendSearchResults[i].m_RoomExtDataReceived)
 			free(m_aFriendSearchResults[i].m_RoomExtDataReceived);
@@ -3358,7 +3358,7 @@ void SQRNetworkManager_Orbis::ProcessSignallingEvent(SceNpMatching2ContextId ctx
 void SQRNetworkManager_Orbis::SignallingEventsTick()
 {
 	EnterCriticalSection(&m_signallingEventListCS);
-	for(int i=0;i<m_signallingEventList.size(); i++)
+	for(size_t i=0;i<m_signallingEventList.size(); i++)
 	{
 		SignallingEvent& ev = m_signallingEventList[i];
 		ProcessSignallingEvent(ev.ctxId, ev.roomId, ev.peerMemberId, ev.event, ev.error_code);
