+++ Minecraft.Client/PSVita/Network/SQRNetworkManager_Vita.cpp
@@ -608,7 +608,7 @@ bool SQRNetworkManager_Vita::FriendRoomManagerSearch()
 	}
 
 	// Free up any external data that we received from the previous search
-	for( int i = 0; i < m_aFriendSearchResults.size(); i++ )
+	for( size_t i = 0; i < m_aFriendSearchResults.size(); i++ )
 	{
 		if(m_aFriendSearchResults[i].m_RoomExtDataReceived)
 			free(m_aFriendSearchResults[i].m_RoomExtDataReceived);
